backend/requirements.txt
boto3>=1.35.0
botocore>=1.35.0
pydantic==2.5.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncio==3.4.3
typing-extensions==4.8.0
dataclasses-json==0.6.3
mangum==0.17.0
requests==2.31.0